Add new test cases for the all_gather CCL op (#2372)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2368

### Problem description
We have only basic silicon tests with all_gather operation.

### What's changed
Add new test cases for the all_gather CCL op
  - Test with various all_gather_dim attributes.
  - Test with various input tensor ranks.
  - Test with increased input tensor dimensions.
  - Test with both fp32 and bf16.

Tests with tensors that are not multiples of the tile size were not
included. They cannot be run with the runtime due to this issue:
https://github.com/tenstorrent/tt-metal/issues/16870. These testcases
will be updated when the runtime issue resolved.

_All tests executed successfully, and their outputs matched the golden
results from torch distributed._

### Checklist
- [x] New/Existing tests provide coverage for changes