Implement runtime tensor desc API (#2370)
This change implements an API on runtime tensors to reflect tensor
metadata and contents reflecting the underlying TTNN tensor. This
functionality is pybound as well, allowing for easy casting to torch
tensors. Please see `runtime/test/python/ttnn/test_runtime_api.py` for
an example.

NOTE: This is not the most efficient implementation possible, as there
is effectively a double copy to get the data in a pybindable row major
format. There is probably some trickery that can be done later to avoid
this, but I would like to get this functionality out ASAP and avoid
premature optimization

Big thanks to @tapspatel for helping me figure out the namespace
dispatch intricacies here!

### Ticket
Closes #1957 

### Checklist
- [x] New/Existing tests provide coverage for changes