#2403: Updated default system descriptor attributes in compiler to handle multi-device meshes (#2411)
This change adds support for multi-device default values for the system
descriptor in the compiler. It assumes a linear mesh connected in ring
topology connections. This does not map to a physical layout but allows
the user to run compiler passes on a multi-device configuration without
having to provide a multi-device system descriptor.

With change, you can now do this
```
ttmlir-opt --ttir-to-ttnn-backend-pipeline="mesh-shape=2,4" test.mlir
```