Change Manual shard_type to Identity and remove unnecessary mesh_shar… (#2373)
…d op

### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2337
Part of https://github.com/tenstorrent/tt-mlir/issues/2367

### Problem description
1. Current description of shard_type is mainly from stablehlo notations.
In particular, "Manual" mode in shard_type of mesh_shard op provides
some confusion in interpreting the corresponding sharded tensor. It
basically indicates that the input and output shards are identical such
that we don't need any further sharding operation. To clarify the term,
we need to update the "Manual" shard type.
2. Current mesh_shard ops are costly in that the operations should
happen on system memory, so we would like to minimize the usage.

### What's changed
1. Replace "Manual" with "Identity" as it better describes the sharded
tensor state.
2. If frontend provides pre-sharded tensor and expects sharded return,
(1) check the sharding attribute of input and return and (2) create
mesh_shard op only if it is necessary to maintain shape matching
purpose.
3. Relaxed system memory requirements for the dummy mesh_shard ops to
avoid copying from and to device memory.

### Checklist
- [X] New/Existing tests provide coverage for changes