EmitCTTNNEmitter refactor (#2457)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2455

### Problem description
Order of SourceTy and TargetTy is flipped at the moment, so explicit
instantiation with just TargetTy is impossible (which goes against
original intention).
### What's changed
Flipped the order of `SourceTy` and `TargetTy` in `emit` member
function. Added a `TTNNTarget` template struct that maps type to the
only TTNN C++ type that would make sense, so we can still instantiate a
lot of cases implicitly.

### Checklist
- [x] New/Existing tests provide coverage for changes