Add TTIR pass that lowers DMA ops from affine to index form
This pass lowers DMA ops from their affine form to indexed form. This is useful for doing analysis on the DMA
ops and lowering them to an optimal loop nest of coalesced transactions.  This is acheived by sampling the affine
map over the entire parent generic op iterator space. Note that the affine map provided to the DMA op must be
one of the indexing maps of the parent generic op.

e.g.
```mlir
%tx = ttir.dma %stream[#map1], %cb0
```

Might become:
```mlir
%c2 = arith.constant 2
%iter0 = ttir.iter_index(0)
%core0 = ttir.core_index(0)
%0 = arith.muli %core0, %c2
%1 = arith.addi %0, %iter0
%iter2 = ttir.iter_index(2)
%tx = ttir.dma %stream [%1, %iter2], %cb0
```