Refactor tt.device as module symbol and optional in tests (#2487)
The primary motivation for this change is to simplify tests. Today many
tests have system descriptors and device attributes copied directly in
the IR, this is fragile and annoying when it comes to updating system
descriptors or the device attribute. Secondary motivation is some code
cleanup and simplification in the creation of and working with these
attributes.

This commit implements a few relatively simple cleanups of device and
system desc creation. Unfortunately due to the nature of these core
components the change turns out to be relatively large, though mostly
mechanical. These changes include:

- `--ttir-load-system-desc` and `--ttir-implicit-device` are now merged
under a single pass called `--tt-register-device`. The pass has also
correctly moved to reside in the TT dialect. `--ttir-load-system-desc`
is kept for now as a deprecated pass so as not to break too many offline
scripts.
- `TTDevice` type has now moved to TTNN dialect and no longer references
the top level attribute directly so that a device (and transitively a
system desc) doesn't have to be manually declared in tests.
- No more forceReload on the placeholder device attribute.
- The device is also a top level op that lives in the module's symbol
table. This better models the scope and object of the device. It also
uses MLIR's symbol table machinery, which simplifies device lookup.
- Many test updates to adopt these changes

What used to be the module level attribute `tt.device` is now a named op
in the module's symbol table (named `default_device`), it looks like:
```mlir
module {
  tt.device @default_device = <workerGrid = #tt.grid<8x8, (d0, d1) -> (0, d0, d1)>, l1Map ...>
  ...
}
```
- Note this could open the door for having multiple devices defined at
the same time or reinterpretations of the same system or sub-systems. It
might be required for pipeline parallelism.

In TTNN Dialect, what used to read:
```mlir
%0 = "ttnn.get_device"() <{mesh_shape = #ttnn<mesh_shape 1x1>}> : () -> !tt.device<#device>
%2 = "ttnn.to_device"(%1, %0) : (tensor<32x32xbf16, !tt.device<#device>) -> tensor<32x32xbf16>
```

Is now (note no longer references device attr, so no longer required to
device this attribute in tests):
```mlir
%0 = "ttnn.get_device"() <{mesh_shape = #ttnn<mesh_shape 1x1>}> : () -> !ttnn.device
%2 = "ttnn.to_device"(%1, %0) : (tensor<32x32xbf16, !ttnn.device) -> tensor<32x32xbf16>
```