#2506: Fixed bug where flatbuffer dependencies were not getting regenerated in incremental build (#2517)
When trying to do incremental builds, there was a bug where flatc
compiler did not regenerate the correct dependencies for a root
flatbuffer schema if any of the dependencies were changed. This lead to
seg faults when trying to read the flatbuffer in ttrt (since the runtime
was using the serialized schema bytes from an old build and not the
latest).