Bufferization tensor shape normalization (#2490)
Bufferization has a weird constraint that the tensor -> buffer
conversion must maintain the same shape and rank. Since we use the
tensor encoding to express shape collapse this becomes problematic
because the tensor's shape will rarely match that of the collapsed
shape. This primary function of this change is to add a tensor shape
normalization pass, such that all tensors inherit the shape of their
layout encoding so that it can 1-1 match the bufferized memref. This new
pass makes it is possible to then perform bufferization.

New bufferization related passes and setup has been moved to
Bufferization.cpp. I also took this as an opportunity to commonize
`TTIRTensorTypeRewriter` since it is now used by 3 different passes and
is a nice utility.