[runtime] add option for enabling program cache
Add flag to `openDevice()` so that the program cache can be enabled.

I didn't want to add one more parameter to the list, so there is a bit
of refactor. `OpenDeviceOptions` struct is introduced to encapsulate all
of the exposed options.

This has potential to break frontends, from a quick glance, it looks
like `tt-torch` is using the async flag, `tt-xla` and `tt-forge-fe`
should be fine.

Issue tenstorrent/tt-forge-fe#1493