[runtime] add option for enabling program cache
Add flag to `openDevice()` so that the program cache can be enabled.

Initially wanted to refactor the `openDevice()` so that it uses a struct
to encapsulate all of the open device options. However, we've decided to
add `enableProgramCache` to the function parameter list for now. There is
some other refactoring work in progress around opening the device and
changing the API, so leaving the cleanup for later.

Simple unit test for verifying wheter the program cache is
enabled/disabled is added.

Issue tenstorrent/tt-forge-fe#1493