clean up createTTIRToTTMetalBackendPipeline() and related (#2630)
### Problem description
We've discussed cleaning up the flow in
`createTTIRToTTMetalBackendPipeline()` given that we no longer have a
viable v0 pipeline in `main`. A couple of pass options for switching b/w
v0 and v1 have been removed and the corresponding pipeline and passes
have been cleaned up accordingly.

This also takes care of a code TODO in
`createTTIRToTTMetalBackendPipeline()` to insert the pass step for
`ttir.generic` conversion.

### What's changed
 - rm TTIRToTTMetalBackendPipelineOptions.version
- rm TTIRAttachMetalLayout.useStreamLayout option ('true' is now v1
default)
 - simplify conversion flow in AttachMetalLayout correspondingly
- sandwich createTTIRGeneralizeNamedOps() after the AttachMetalLayout
pass in createTTIRToTTMetalBackendPipeline()
 - update one impact lit test

### Checklist
- [x] test/ttmlir/Dialect/TTIR/generalization/generalization.mlir
updated to no longer using 'useMetalLayout' option