Bug fix, generate semaphores for all generic regions in lockstep (#2533)
This commit actually fixes a couple of issues in the semaphore
generation part of the datamovement generation pass. These include:

- Pregenerating all of the region function signatures for the new
generic op, up front. This is the core bug, previously we were just
creating new regions on the fly, but this meant that they would miss out
on previous region's semaphore additions.
- Compute region should also inherit semaphores.
- Modify only the newly created regions instead of the old op in place.
- Additional generic op verification and testing to enforce that all
regions have the same number of arguments and the same type for each
respective argument.

Closes #2526