Refactor `zeros` and `ones` under one op class (#2619)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2584

### Problem description
`zeros` and `ones` both naturally and implementationally in TTNN share
the same structure. This means that treating them separately is just an
unnecessary code duplication.

### What's changed
They are now treated as one class both in TTIR and TTNN dialects, as
well as in FB and runtime. TTNN uses `FullWith<T>` as a struct name for
these ops, I have chosen `NamedFull` in dialects and FB.

Limitations: I treated these changes as pure refactoring, so there might
be some things that need additional treatment that are not covered by
this PR.

### Checklist
- [x] New/Existing tests provide coverage for changes