Add a new pass that checks the match of input and output of tensor (#2676)
annotations and fix if there is any issue.

### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2645

### Problem description
Found cases where stablehlo from JAX may not fully follow sharding rules
for automatic parallelism with sharded inputs/outputs.

### What's changed
Instead of fixing the issue with case by case, the new pass is
introduced that allows us to manage tensor annotations in a more general
manner. In particular, this pass address following two cases.
1. Single device tensor op has multi device tensor argument in a single
op.
In this case, we need to insert MeshShardOp before the op and change the
 tensor to single device tensor.
2. Multi device tensor ops have missing multi device tensor annotation
in
case of automatic parallelism with pre-sharded inputs and outputs. In
this
 case, we need to add missing TensorMeshShardingAttr.

### Checklist
- [X] New/Existing tests provide coverage for changes