ci_compile_dylib script improvements (#2729)
### Ticket
Closes https://github.com/tenstorrent/tt-mlir/issues/2095

### Problem description
There are a couple of problems in the ci_compile_dylib.py script
described in the above issue.

### What's changed
This PR adds the ability to specify a different build directory and
provides an option to compile a single C++ file instead of a whole
group. It also polishes the MD file on how to run the emitC tests
locally.

### Checklist
- [ ] New/Existing tests provide coverage for changes