Insert deallocation ops for func op input params, track tensor lifetime in runtime (#2818)
### Ticket
closes #2459

### Problem description
Currently we do not deallocate func op inputs within graph execution,
even when these func op inputs are no longer used. Since these arguments
are provided by the user, we need a way to determine the lifetime of the
inputs and whether or not they can be deallocated.

One way to go about this would be to base it off the input argument
type. However there are cases where we would want to deallocate
constants as well.

Another way would be to add an additional attribute to indicate whether
an input can be deallocated. This would require recompile every time the
lifetime of a tensor changes at runtime.

Therefore we decided it makes more sense to configure input lifetime at
runtime rather than compile time.

### What's changed
* Updated dealloc pass to deallocate func op inputs
* Updated ttnn runtime to track a `retain` flag that is configurable by
the user. Deallocate operations within the program will not deallocate
the tensor if the `retain` flag is set to `true`
   * The retain flag is defaulted to `true`
* Added tests to make sure that tensors are retained/deallocated as
expected.

An example would looks something like:
```
for constant in constants:
   runtime.setRetain(constant, True)
   
for activation in activations:
   runtime.setRetain(activation, False)

# Tensors in `activations` will be deallocated within the submit run once they have no more users
# While tensors in `constants` will remain allocated
outputs = runtime.submit(..., activations + constants)
```
### Checklist
- [X] New/Existing tests provide coverage for changes
- [ ] FE updates should be simple. Since `retain` defaults to `true`, no
inputs will be deallocated by default, thus the overall behaviour is the
same as before. The only difference is that submit now takes
`std::vector<Tensor> &inputs` instead of `const std::vector<Tensor>
&inputs`, and so the FE bindiings may need updates to remove the `const`
qualifier.