Use LLVM provided STL utils from future versions (#2839)
### Ticket
N/A

### Problem description
LLVM currently builds with C++17 but provides some STL utilities that
we've been using that are available in future versions
https://llvm.org/doxygen/STLForwardCompat_8h_source.html.

### What's changed
Replaced existing functionalities with provided utilities, this will
also be useful once LLVM starts to build with C++20/23, as these
utilities will be marked as deprecated, and it will be easier to replace
them with STL implementation. Fixed includes in touched files.

### Checklist
- [ ] New/Existing tests provide coverage for changes