Update runtime API for tensor creation (#2842)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2757

### Problem description
Explained in details in the ticket
https://github.com/tenstorrent/tt-mlir/issues/2757

### What's changed
- Changed tensor creation API functions to use `void *` instead of
`shared_ptr` so frontends don't have to create `shared_ptr` with empty
deleters to use the runtime API.
- Renamed tensor creation API functions to state clearly on which device
is buffer created and with which storage type, and added documentation
comments in header files. Beside more clarity about what each function
does, now it's much easier to track usages of each function across the
codebase, before if you searched `createTensor` you'd get tons of hits
and would have to look at argument types.
- Added new tensor creation API function to create multi-device tensor
from already existing tensor shards. We need this in tt-xla because we
first create tensors for each input buffer (that's how PJRT API works),
and then after compilation when we know the multi-device strategy we
need to create multi-device tensor (if needed), so we need to create it
from already existing tensors.

### Checklist
- [x] New/Existing tests provide coverage for changes

I've prepared uplift PRs for frontends:
- tt-torch: https://github.com/tenstorrent/tt-torch/pull/568
- tt-forge-fe: https://github.com/tenstorrent/tt-forge-fe/pull/1707
- tt-xla: I'm making on some other changes there that will make use of
this, so I will send PR for both later

Frontend on-pr CI jobs passed:
-
[tt-torch](https://github.com/tenstorrent/tt-torch/actions/runs/14285792087)
-
[tt-forge-fe](https://github.com/tenstorrent/tt-forge-fe/actions/runs/14286335280)