TT-MLIR logger
Added simple logger for tt-mlir compiler.
It is used with ttmlir-opt when option -debug is set.
Further we can filter with
-debug-only="<log-component1>,<log-component2>..."

It works only in debug build as it is meant for dev purposes.
Env variable TTMLIR_LOGGER_LEVEL can be set to set log level to
TRACE which should be the most verbose. Default level is DEBUG
which will always be printed.

TTMLIR_ENABLE_DEBUG_LOGS is the compile time flag used
to control macro expansion so in non-debug builds log statements
will be void.