Add initial support for PrepareConv2dWeights op (#2828)
### Ticket
https://github.com/issues/assigned?issue=tenstorrent%7Ctt-mlir%7C2636

### What's changed
Added initial support for the `PrepareConv2dWeights` op. It is currently
not fully supported because we can’t calculate the shape and layout of
the output tensor from this op—there is a lot of logic happening in
Metal we would have to copy. We will call this op through graph capture
at compile time in the optimizer to determine the shape and layout after
Metal has returned the tensor.