Erase Inverse Ops pass (#2528)
### Problem description
This pass aims to purge the graph of TMs which are the inverses of each
other when separated by a sequence of other ops.

### What's changed
- Added `EraseInverseOps` pass
- This pass applies _commute patterns_ to ops in the graph. That is the
pattern will match some op, followed by a TM and move the TM to the
producer of the op
- Patterns have been implemented to commute transposes, reshapes, and
permutations through eltwise unary, eltwise binary, and broadcast ops.

- Added/replaced TM canonicalization patterns using `fold`
- This is because folding patterns run **during** passes automatically,
whilst canonicalization must be explicitly called.
- These patterns are responsible for erasing back-to-back inverse TMs.
The folding patterns are:
        - Replace identity TM with its operand
- For back-to-back permutations, fuse the permutation attribute into the
consumer permutation, and set its operand to the input of the first
permutation
- For back-to-back inverse transposes, replace the consumer with the
producer transpose's input
        - Replace back-to-back reshape with just the consumer reshape
- While these patterns do not explicitly remove inverses, they do after
multiple calls to `fold`. I.e, back to back permutation will become just
one permutation, that permutation ends up being the identity
permutation, then that permutation is removed.


- Added TTIRToTTNNBackendPipeline option to enable/disable the pass
`enable-erase-inverse-ops-pass` (`true` by default)
- Added `FlattenSlidingWindow` pass and `FlattenedCompatInfoAttr`
- Since ttnn sliding window ops require their inputs (and outputs) be in
the format `(1, 1, N*H*W, C)` we need to insert reshapes on the
input/output of these ops at some point during lowering.
    - Currently this is done when converting TTIR to TTNN
- `EraseInverseOps` has the capability of removing inverse reshapes.
Since it operates on the TTIR dialect we would like to expose these
reshapes before this pass
- `FlattenedCompatInfoAttr` is an optional attribute for `ttir.conv2d`
and `ttir.max_pool2d` and it contains any information that is lost when
flattening the input and output
        - Input height
        - input width
        - batch size
- The `FlattenSlidingWindow` pass inserts the required reshapes on the
input and output of `ttir.conv2d` and `ttir.max_pool2d` and populates
the `FlattenedCompatInfoAttr`
- Lowering `ttir.conv2d` and `ttir.max_pool2d` now requires that
`FlattenedCompatInfoAttr` is populated. That is, `FlattenSlidingWindow`
must be run on the TTIR graph before dialect conversion

- The EraseInverseOps pass has been written for the TTIR dialect. This
is because:
- This is a high level optimization. That is, this is hardware agnostic
- it solely removes high level operations from the program
- We want to do this before memory modelling as it will change the
shapes and attributes of some ops, and in some cases move TMs elsewhere
in the graph (if they are not erased)
- There is also the direct to metal backend which would likely want to
use this pass during lowering too
 
Transposes/Permutes/Reshapes can conditionally commute through ops like
Reduce, Select, and Concat. The commute logic for those has not been
implemented. We can add this logic at any time by defining new a
`CommuteRewritePattern`.
- The `CommuteRewritePattern` classes for eltwise unary and binary ops
are currently identical. However I've separated them as we may wish to
implement different logic in `isCommuteFavorable` given that eltwise
binary ops have two operands.

There is room for a more rigorous analysis on whether we should commute.
Currently we only commute if all users of an eltwise op are an identical
TM, however there can be cases where you want to commute anyway. The
`isCommuteFavorable` functions can be as detailed as we like.

### Checklist
- [X] New/Existing tests provide coverage for changes