#2450 Add an API to convert a TensorSpec into a TTNNLayoutAttr (#2873)
### Ticket
#2450

### Problem description
Searching for legal output layouts takes a long time in the optimizer
since most ops only allow one or a few output layouts for a given input
layout. We should let the op set the output and simply read it back. As
of https://github.com/tenstorrent/tt-metal/issues/19198 we have the
output `tensorSpec` available

### What's changed
- Create an API that takes the returned `tensorSpec` and converts it to
a `TTNNLayoutAttr`
- This requires separately converting every field in a `tensorSpec` to
the corresponding type in MLIR, then creating a new `TTNNLayoutAttr`.
This would be done in the same `context` as the provided input/output
layouts
- Integration into OpModel APIs will be done in a separate PR to keep
the reviews manageable
- Add unit tests 

### Checklist
- [X] New/Existing tests provide coverage for changes