TT-MLIR logger (#2884)
### Problem description

Currently there is no proper way to log things around in tt-mlir
compiler (except in runtime which has its own logger). We can't reuse
the same logger as it is built specifically for runtime with exceptions
which are not supported in llvm based tt-mlir.

### What's changed

LLVM provides with basic macro for debugging. This commit adds a logger
which is basically a wrapper around it. It will be possible to specify
log level (verbosity) and log component (optimizer, d2m, ttnn
dialect...). To run with logs enabled we have to run in debug build, and
set option `-debug` to enable all logs. This includes mlir logs so
better use would be to set option `-debug-only=<log-component>`.

Log looks like:
<img width="669" alt="image"
src="https://github.com/user-attachments/assets/116e7299-2cb1-46c2-bfc7-846a2f9480db"
/>


### Checklist
- [x] New/Existing tests provide coverage for changes