Don't fold during element type normalization (#2932)
Given this IR:
```mlir
  %0 = ttir.empty() : tensor<1x256x1xf64>
  %1 = "ttir.broadcast"(%arg0, %0) <{broadcast_dimensions = array<i64: 1, 1, 1>}> : (tensor<1x256x1xf64>, tensor<1x256x1xf64>) -> tensor<1x256x1xf64>
  return %1 : tensor<1x256x1xf64>
```

When we are updating types we start from `return` op and we update it's
first operand to `f32`. This will prompt mlir to update `ttir.broadcast`
return type to `f32` (probably to keep IR consistent) , so after the
update on `return` op we end up with:

```mlir
  %0 = ttir.empty() : tensor<1x256x1xf64>
  %1 = "ttir.broadcast"(%arg0, %0) <{broadcast_dimensions = array<i64: 1, 1, 1>}> : (tensor<1x256x1xf32>, tensor<1x256x1xf64>) -> tensor<1x256x1xf64>
  return %1 : tensor<1x256x1xf32>
```

Then next up is broadcast op. `applyPatternsGreedily` first try to fold
op it wants to pattern match to avoid unnecessary rewriters, so in case
of our broadcast its foldable and we return `%arg0`, but this is invalid
since fold logic expects fold result to be of the same type as operation
result. In this case `%arg0` is f64 but `ttir.broadcast` is `f32`.

As simple fix I've added canonicalization as first pass in pipeline,
this solves the issue. We could also turn off folding in pipeline
itself, but that might elide opportunity for some folds but I'm not
against turning it off in pipeline also.