[EmitC] Add TTIR to EmitC SO pipeline (#2962)
### Ticket
#2953 

### Problem description
TTNN to EmitC conversion is done in 2 ways: one for ttnn-standalone, and
one for ttnn-dylib (i.e. shared object path).

We already have a pipeline for ttnn-standalone (ttir-to-emitc-pipeline),
we want to add ttir-to-emitc-so-pipeline for the shared object path.

These pipelines differ only in surrounding boilerplate around the
emitted forward function. The ttir-to-emitc-pipeline adds a function to
generate inputs, and a main fn to drive execution, while the
ttir-to-emitc-so-pipeline will repackage forward function to take in a
vector of tensors and a device (targeted for shared object compilation).

### What's changed
Added the new pipeline, added a test, changes to
`ttir-to-emitc-pipeline` test making it more robust.

### Checklist
- [x] New/Existing tests provide coverage for changes