[Optimizer] Integrate TensorSpec to TTNNLayoutAttr conversion into getOpConstraints and return the TTNNLayoutAttr to the caller (#2935)
### Ticket
#2450

### Problem description
Searching for legal output layouts takes a long time in the optimizer
since most ops only allow one or a few output layouts for a given input
layout. We should let the op set the output and simply read it back.

As of #2873 we have the conversion API required for getting a
`TTNNLayoutAttr` from the `TensorSpec` returned from the metal API

### What's changed
- Integrate the conversion API with `getOpConstraints()`
- Pass in the device grid from the operation and the `MLIRContext` to
the MLIR-Metal boundary
- After receiving the constraints response from Metal, call the
conversion function to construct a `TTNNLayoutAttr` from the
`TensorSpec`
  - Return the `TTNNLayoutAttr` all the way back to the callers
- Ensure `getOpConstraints()` can be called with `outputLayout` set to
`nullptr`
- Added tests for conv2d, matmul, add, and relu to do sanity checks on
the returned `outputLayout`
- Update unit tests and existing callers
- Enabled conv2d unit tests
- Closes #2450

### Checklist
- [x] New/Existing tests provide coverage for changes