Introduce new workaround for all_reduce op. (#2764)
### Ticket
Closes #2675 

### Problem description
We decompose all_reduce into reduce_scatter and all_gather as a
workaround. However, reduce_scatter requires that the size along
scatter_dim be divisible by the number of devices. This constraint
propagates back to all_reduce, limiting our ability to compile certain
models. In cases where this divisibility requirement is not met,
compilation can fail.

### What's changed
Added a new decomposition path for the AllReduce TTIR op using AllGather
+ local Reduce.
This serves as an alternative to the existing ReduceScatter + AllGather
workaround.
The new decomposition is applied when:
- The tensor shape along the scatter_dim is not divisible by the number
of devices, or
 - The memory overhead is smaller than heuristically found size_limit

### Checklist
- [x] New/Existing tests provide coverage for changes