CumSumOp negative axis index attr (#3062)
### Ticket
Part of the https://github.com/tenstorrent/tt-mlir/issues/3053

### Problem description
CumSumOp assumed that axis index attribute will be in the range [0,
rank).

### What's changed
Relaxed the condition so it now accepts axis index in the range [-rank,
rank) on the input and normalize that range to [0, rank) as a part of
the folding.

### Checklist
- [x] New/Existing tests provide coverage for changes