Make all params or `Conv2dConfigAttr` optional (#2912)
If we want to fuse relu and conv we need to set activation param in
`Conv2dConfigAttr`, but since majority of it's params are non-optional
we would have to determine correct parameters. This PR changes all
params of `Conv2dConfgAttr` so that we can easily specify which
parameter we want to override and leave the rest to default to whatever
is set in ttnn.