Add some ttir.gather -> ttir.embedding cases (#2812)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-mlir/issues/1350)
(Not all cases are covered)

### Problem description
`ttir.gather` that is lowered to `ttir.embedding` in TTIR to TTIR
decomposition was implemented with constraints that can be loosened up,
with more cases covered.

### What's changed
-`ttir.embedding` supports indexing the first dimension of a 2D tensor,
and getting slices of full size from the second dimension. We call this
tensor `weight`.
-The other operand is a tensor that has the start indices. This tensor
can be 1D or 2D and we call it `input`.
-Output shape is the shape of input with the last dimension of weights
appended (indices are replaced with corresponding slices from the
weights).

Cases of `ttir.gather` that will be supported after this PR are those
where:
-The dimensions we are indexing have slice size 1 and don't appear in
the output as offset dimensions -> collapsedSliceDims are the same as
startIndexMap
-For all other dimensions, slice size has to be equal to dimension size
(because embedding takes the whole row) -> sliceSizes = [1, 1,
...fullDim, fullDim, ...]
=> This mean we can permute and reshape the operand appropriately. First
we permute so all of the indexing dims are at the start of the operand.
We then take the first few dimensions that we are indexing and flatten
them together to be the first dimension. We also flatten the other
dimensions together to be the second dimension.
This 2D tensor is `weight` for `ttir.embedding`.

-There are no input or start indices batching dimensions, as this can't
be lowered directly to `ttir.embedding`.

-start indices tensor can't be 1D when we are indexing multiple
dimensions. This is because `ttir.matmul` that transforms start indices
needs a 2D tensor.
=> This means we can reshape start indices tensor appropriately. If it
is already rank 1 or 2, we leave it as to not add unnecessary reshapes.
Otherwise we flatten the dimensions together.
=> If more than one dimension of operand is being indexed, we add
`ttir.matmul` to transform the indices to match the flattened operand.
This requires a permutation if index vector dim isn't the last
dimension.
This 1D or 2D tensor is `input` for `ttir.embedding`.

-Offset dimensions can be any length. Output shape of 'ttir.embedding'
It will always have the slices as last dimension (and offset dimensions
represent where the slices are), so we need a permute op if offset
dimensions aren't the last dimensions of the output.
=> If necessary, output will be reshaped as well, to match `ttir.gather`
output. This will be the case when we need to restore dimensions that
were flattened.

### Checklist
- [x] New/Existing tests provide coverage for changes