[TTNN] Introduce TTNNPrepareConv2dWeights pass (#2897)
### Ticket
[Link to Github
Issue](https://github.com/issues/assigned?issue=tenstorrent%7Ctt-mlir%7C2636)

### Problem description
Each time the `ttnn.conv2d` operation is executed, it preprocesses the
weights under the hood. This introduces significant performance
overhead, as a lot of time is spent preparing the weights on the host
and then copying them to the device.

To address this, instead of allowing `ttnn.conv2d` to implicitly prepare
the weights each time, we’ve extracted a dedicated operation called
`ttnn.prepare_conv2d_weights`. The idea is to insert this operation
before every `ttnn.conv2d` in the graph so that it can be consteval’d
once consteval support is complete.

### What's changed
Introduced the `TTNNPrepareConv2dWeights` pass, which inserts
`ttnn.prepare_conv2d_weights` before every `ttnn.conv2d` op in the
graph.

When calculating the output tensor’s shape and layout from
`ttnn.prepare_conv2d_weights`, we currently rely on a graph capture API
created by @odjuricicTT. This API simulates the behavior of
`ttnn.prepare_conv2d_weights`, allowing us to retrieve the shape and
layout.

In the next iteration, we plan to use an API that Metal needs to expose,
which will directly compute the shape and layout. This change is
important, as the current approach impacts compile time and is not very
clean.