Create DeviceLayoutInterface (#3096)
This interface implies a MemRefLayoutAttrInterface that conforms to
Tenstorrent device layout. Tenstorrent devices are divided into a grid
of cores, onto which we map tensors that have been chunked into shards.
This device layout attribute semantically means exactly this, i.e. we
have some tensors that are distributed amongst the device core grid.
This means that memref shapes should be reinterpreted as a concatenation
of a grid shape and a shard shape, for example, memref<1x2x3x4xtt.tile>
should be interpreted as a grid of 1x2 and a shard of 3x4. This
interface provides some convenience methods for working with this
interpretation of memref shape.

This commit introduces this interface and fixes up all of the places
where we previously assumed this layout.