[Optimizer] Sharding transpose op and new layout analysis (#2982)
### Ticket

Related to https://github.com/tenstorrent/tt-mlir/issues/2637
Closes https://github.com/tenstorrent/tt-mlir/issues/2911
Closes
https://github.com/orgs/tenstorrent/projects/54/views/8?sliceBy%5Bvalue%5D=Optimizer&pane=issue&itemId=98707119&issue=tenstorrent%7Ctt-mlir%7C2219


## What's changed

### **Sharding transpose op**

* Enable sharding of transpose op with row-major layout because that is
what works currently in TTNN. Tranpose is a complex op and its sharding
requirements depend on dims that are tranposed. In resnet case we're
transposing HC dims (NCHW format) which requires row-major input.
* Resharding when first op in graph does not support interleaved input.
This is covered by sharding tranpose op test.
* When generating layouts do not generate multiple grid solutions for
the same shard shape. This is done per shard strategy, i.e. we can have
same shard shape with H, W or HW sharding.
* Fix toLayout op handling in decomposition pass when tensor bounces to
host and back. Bring back output tensor layout's grid.

### **Split layout analysis**

* Introduced new AllPossibleLayoutsAnalysis that generates them.
* Also, another analysis at the begin of pipeline is to pick up all
scalar types for which we generate layouts called
ScalarDataTypeAnalysis.
* LegalLayoutAnalysis now picks up already generated layouts. It just
takes layouts that are generated for the wanted ranked tensor of the op
for which it is running.
* ShardSolver receives map of all possible layouts for every tensor in
the graph and uses it when it needs to create reshard op.
* Unit test for sharding tranpose op with override that allows
row-major.
* Unit test for new analysis for layout geenration.


### Checklist
- [x] New/Existing tests provide coverage for changes