Remove redundant context param from `with*` attribute member functions (#3073)
### Ticket
N/A

### Problem description
`with*` member functions of `MemoryConfigAttr` and `TTNNLayoutAttr` are
taking `MLIRContext` as a first parameter. The idea behind these
functions is to create the attribute that's the same as the provided one
except for one of its parameters changes. I can't think of any case were
we would want to instantiate that attribute in the different context
than the provided one, but if someone can think about the use-case
please let me know.

### What's changed
Removed `MLIRContext` param from those functions.

### Checklist
- [x] New/Existing tests provide coverage for changes