Add workaround for int broadcasting in 'Add' op
TTNN currently fails to handle broadcasting for int32 inputs in the `add` op,
which is used internally to represent `expand` operations. When an `expand` op
with int32 is lowered to `add` with broadcasting, execution does not work as expected.

As a temporary workaround, inputs are converted to bfloat16 when shapes differ.
This should be removed once TTNN properly supports broadcasting for int32.